implemented multiple plexus.xml support. plexus.xml is now treated in much the same way as components.xml, except for the extraction of container services (componentFactoryManager, etc.)...the plexus.xml used for the container is built up from all available plexus.xml files in the startup classpath.
Also, DefaultComponentFactoryManager now looks up factories using the container, to allow dynamic injection of new component factories. It's hardwired to return a certain factory for java stuff, tho...just to be safe, and to avoid a recursive search for the java factory by looking up the java factory. :)

Added tests to (partially, not satisfactorily) test all of this.